+++ docs/changes.rst
@@ -15,6 +15,7 @@ Not yet released.
 * Track comment removal/resolving in history.
 * Fixed restoring project backups with linked components.
 * Fixed captcha entering on unsuccessful registration.
+* Improved languages support in DeepL.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/87?closed=1>`__.